backend/requirements.txt
# FastAPI ve Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1

# AI ve Image Processing
torch==2.5.1
torchvision==0.20.1
opencv-python==4.12.0.88
pillow==11.1.0
numpy==2.2.6

# GFPGAN ve RealESRGAN
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
realesrgan==0.3.0

# Transformers ve Modeller
transformers==4.57.3
timm==1.0.15

# Utilities
requests==2.32.5
tqdm==4.67.1
yapf==0.43.0

# Development (Optional)
pytest==8.3.4
httpx==0.28.2
